Refactor operations catalog into dedicated route
