+++ app/page.tsx
@@ -1,38 +1,20 @@
 "use client"
 
-import { useEffect, useState } from "react"
+import { useEffect } from "react"
 import { useRouter } from "next/navigation"
-import { useSupabaseClient, useUser } from "@supabase/auth-helpers-react"
-
-import OpsCatalog from "@/components/ops-catalog"
-import { DashboardShell } from "@/components/dashboard-shell"
+import { useUser } from "@supabase/auth-helpers-react"
 
 export default function Home() {
-  const [query, setQuery] = useState("")
   const router = useRouter()
-  const supabase = useSupabaseClient()
   const user = useUser()
 
-  // ✅ redirect if no user
   useEffect(() => {
     if (user === null) {
       router.replace("/login")
+    } else if (user) {
+      router.replace("/operations")
     }
   }, [user, router])
 
-  if (!user) {
-    return <p>Redirecting to login…</p>
-  }
-
-  return (
-    <DashboardShell
-      search={{
-        value: query,
-        onChange: setQuery,
-        placeholder: "Search categories or Processes…",
-      }}
-    >
-      <OpsCatalog query={query} />
-    </DashboardShell>
-  )
+  return <p>Redirecting…</p>
 }
