+++ components/operations/index.ts
@@ -1,0 +1,1 @@
+export { default as OperationsCatalog } from "./operations-catalog"
